+++ NEWS.md
@@ -1,4 +1,22 @@
-# rSOILWAT2 v6.1.0-9000
+# rSOILWAT2 v6.2.0
+* Simulation output remains the same as the previous version unless
+  relative humidity is calculated from vapor pressure or specific humidity.
+* Update `SOILWAT2` to v8.0.1 which fixed the calculation of relative humidity
+  and the count of days where missing weather was imputed by `"LOCF"`.
+
+## New features
+* `validObject()` method for class `"swInputData"` now includes checks for a
+  valid `"weatherHistory"` object.
+
+## Bugfix
+* `dbW_fixWeather()` now handles data objects with all missing values.
+
+## Changes to interface
+* The inputs of daily specific humidity changed units (`"%"` to `"g kg-1"`)
+  and name (`"specHavg_pct"` to `"specHavg_gPERkg"`).
+
+
+# rSOILWAT2 v6.1.0
 * This version produces the same output as the previous version.
 * `SOILWAT2` updated to v8.0.0 which now includes a simulation domain;
   however, this has no impact on `rSOILWAT2`.